+++ src/app/CMakeLists.txt
@@ -3,14 +3,13 @@ project(app)
 
 add_executable(app main.cpp)
 target_include_directories(app PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../include)
-target_link_libraries(app PRIVATE core)
+target_link_libraries(app PRIVATE engine)
 
 find_package(GLEW REQUIRED)
-find_package(glfw3 CONFIG REQUIRED)
 find_package(imgui CONFIG REQUIRED)
 find_package(SFML COMPONENTS system window graphics CONFIG REQUIRED)
 
-target_link_libraries(app PRIVATE GLEW::GLEW glfw imgui::imgui)
+target_link_libraries(app PRIVATE GLEW::GLEW imgui::imgui)
 target_link_libraries(app PRIVATE sfml-system sfml-network sfml-graphics sfml-window)
 
 file(COPY ${RESOURCE_DIR} DESTINATION ${CMAKE_RUNTIME_OUTPUT_DIRECTORY})
